project/pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Yegor Shevchenko <egorhs19@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
django-rest-framework = "^0.1.0"
celery = "^5.2.1"
flower = "^1.0.0"
redis = "^4.0.2"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
pydocstyle = "^6.1.1"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
